+++ lib/active_delivery.rb
@@ -11,4 +11,5 @@ require "active_delivery/callbacks" if defined?(ActiveSupport)
 require "active_delivery/lines/base"
 require "active_delivery/lines/mailer" if defined?(ActionMailer)
 
+require "active_delivery/raitie" if defined?(::Rails::Railtie)
 require "active_delivery/testing" if ENV["RACK_ENV"] == "test" || ENV["RAILS_ENV"] == "test"